Correction of comments
Test methods were moved from the deleted class to ProductRepositoryTest. Classes and their files were removed to implement console input. Data access to the database was implemented using launch arguments.
