+++ src/main/resources/application.properties
@@ -2,10 +2,10 @@ app.name=NauJavaFoodControl
 server.port=8080
 app.version=0.0.3
 management.endpoints.web.exposure.include=*
-spring.datasource.initialize=false
-#spring.datasource.url=jdbc:postgresql://localhost:5432/...
-#spring.datasource.username=...
-#spring.datasource.password=...
+#spring.datasource.initialize=false
+spring.datasource.url=${DB_URL}
+spring.datasource.username=${DB_USER}
+spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 # JPA
 spring.jpa.hibernate.ddl-auto=update
